move memorycache.js from auth to common folder
