--- app/src/forms/auth/middleware/memoryCache.js
@@ -1,2 +1,0 @@
-const NodeCache = require( "node-cache" );
-module.exports =  new NodeCache();
