+++ app/src/forms/common/middleware/memoryCache.js
@@ -1,0 +1,2 @@
+const NodeCache = require( 'node-cache' );
+module.exports =  new NodeCache();
